refactor(github-project): split project-view and slug-dialog into modules
