+++ src/main/github/project-view.ts
@@ -1,29 +1,34 @@
 /* eslint-disable max-lines -- Why: ProjectV2 GraphQL has its own normalization
-layer, retry policy (parent-field dance), paste-to-add parser, discovery
-pagination, and slug-addressed mutation helpers. Co-locating them keeps the
-retry/classify/normalize contract reviewable as one surface. Splitting would
-risk drifting classification semantics between read and write paths. */
-// Why: `ghExecFileAsync` (WSL-aware, retry-enabled) is the single spawn site
-// for gh calls. The legacy plain `execFileAsync` is NOT used here — routing
-// every gh call through the runner gives us transient-5xx retry, WSL path
-// translation, and a single hook point for future quota tracking.
-import { acquire, release } from './gh-utils'
-import { extractExecError, ghExecFileAsync } from '../git/runner'
-import { rateLimitGuard, noteRateLimitSpend, type RateLimitBucketKind } from './rate-limit'
-import type { GitHubAssignableUser, GitHubWorkItemDetails, PRComment } from '../../shared/types'
+layer, retry policy (parent-field dance), paste-to-add parser, and discovery
+pagination. Co-locating the read path keeps the retry/classify/normalize
+contract reviewable as one surface. Lower-level plumbing (slug validation,
+error classifier, runGraphql/runRest) lives in ./project-view/internals; the
+slug-addressed write path lives in ./project-view/mutations. */
+import {
+  acquire,
+  release,
+  extractExecError,
+  ghExecFileAsync,
+  rateLimitGuard,
+  noteRateLimitSpend,
+  classifyProjectError,
+  driftError,
+  errorsIndicateParentField,
+  rateLimitedError,
+  runGraphql,
+  isValidOwnerSlug,
+  assertSlug,
+  assertPositiveInt,
+  type GhGraphqlErrorShape,
+  type GraphqlVars
+} from './project-view/internals'
 import type {
-  AddIssueCommentBySlugArgs,
-  ClearProjectItemFieldArgs,
-  DeleteIssueCommentBySlugArgs,
   GetProjectViewTableArgs,
   GetProjectViewTableResult,
-  GitHubProjectCommentMutationResult,
   GitHubProjectField,
-  GitHubProjectFieldMutationValue,
   GitHubProjectFieldValue,
   GitHubProjectIteration,
   GitHubProjectLabel,
-  GitHubProjectMutationResult,
   GitHubProjectOwnerType,
   GitHubProjectRow,
   GitHubProjectRowItemType,
@@ -37,24 +42,34 @@ import type {
   GitHubProjectViewLayout,
   GitHubProjectViewSummary,
   ListAccessibleProjectsResult,
-  ListAssignableUsersBySlugArgs,
-  ListAssignableUsersBySlugResult,
-  ListIssueTypesBySlugArgs,
-  ListIssueTypesBySlugResult,
-  ListLabelsBySlugArgs,
-  ListLabelsBySlugResult,
   ListProjectViewsArgs,
   ListProjectViewsResult,
-  ProjectWorkItemDetailsBySlugArgs,
-  ProjectWorkItemDetailsBySlugResult,
   ResolveProjectRefArgs,
-  ResolveProjectRefResult,
-  UpdateIssueBySlugArgs,
-  UpdateIssueCommentBySlugArgs,
-  UpdateIssueTypeBySlugArgs,
-  UpdatePullRequestBySlugArgs,
-  UpdateProjectItemFieldArgs
+  ResolveProjectRefResult
 } from '../../shared/github-project-types'
+
+// Re-export the public API so existing call sites (`./project-view`) keep
+// working unchanged. The split is internal-only.
+export {
+  isValidOwnerSlug,
+  isValidRepoSlug,
+  isValidSlug,
+  classifyProjectError
+} from './project-view/internals'
+export {
+  updateProjectItemFieldValue,
+  clearProjectItemFieldValue,
+  updateIssueBySlug,
+  updatePullRequestBySlug,
+  addIssueCommentBySlug,
+  updateIssueCommentBySlug,
+  deleteIssueCommentBySlug,
+  listLabelsBySlug,
+  listAssignableUsersBySlug,
+  listIssueTypesBySlug,
+  updateIssueTypeBySlug,
+  getWorkItemDetailsBySlug
+} from './project-view/mutations'
 
 // ─── Constants ─────────────────────────────────────────────────────────
 
@@ -98,324 +113,6 @@ export function _resetProjectViewModuleState(): void {
   parentFieldProbeInFlight = null
 }
 
-// ─── Slug validation ──────────────────────────────────────────────────
-
-// Why: GitHub usernames/org logins disallow `_`, `.`, leading `-`. Repo names
-// are looser — they allow leading `_`, `.`, `-` (`.` and `..` reserved). We
-// validate each separately so untrusted Project row data (`nameWithOwner`)
-// can't become an arbitrary REST path while still accepting realistic repo
-// names like `_internal` or `.github`.
-const OWNER_SLUG_RE = /^[A-Za-z0-9][A-Za-z0-9-]*$/
-const REPO_SLUG_RE = /^[A-Za-z0-9._-]+$/
-const REPO_SLUG_RESERVED = new Set(['.', '..'])
-
-export function isValidOwnerSlug(value: unknown): value is string {
-  return typeof value === 'string' && value.length > 0 && OWNER_SLUG_RE.test(value)
-}
-
-export function isValidRepoSlug(value: unknown): value is string {
-  return (
-    typeof value === 'string' &&
-    value.length > 0 &&
-    REPO_SLUG_RE.test(value) &&
-    !REPO_SLUG_RESERVED.has(value)
-  )
-}
-
-// Backwards-compatible alias for callers that don't distinguish owner vs repo.
-// Prefer `isValidOwnerSlug` / `isValidRepoSlug` at new call sites.
-export function isValidSlug(value: unknown): value is string {
-  return isValidOwnerSlug(value) || isValidRepoSlug(value)
-}
-
-function assertSlug(
-  value: unknown,
-  field: 'owner' | 'repo'
-): { ok: true; slug: string } | { ok: false; error: GitHubProjectViewError } {
-  const valid = field === 'owner' ? isValidOwnerSlug(value) : isValidRepoSlug(value)
-  if (!valid) {
-    return {
-      ok: false,
-      error: {
-        type: 'validation_error',
-        message: `Invalid ${field}: "${String(value)}" is not a valid GitHub slug.`
-      }
-    }
-  }
-  return { ok: true, slug: value as string }
-}
-
-function assertPositiveInt(
-  value: unknown,
-  field: string
-): { ok: true; n: number } | { ok: false; error: GitHubProjectViewError } {
-  if (typeof value !== 'number' || !Number.isInteger(value) || value < 1) {
-    return {
-      ok: false,
-      error: {
-        type: 'validation_error',
-        message: `Invalid ${field}: must be a positive integer.`
-      }
-    }
-  }
-  return { ok: true, n: value }
-}
-
-// ─── Error classification ──────────────────────────────────────────────
-
-type GhGraphqlErrorShape = {
-  type?: string
-  message?: string
-  path?: Array<string | number>
-  extensions?: { code?: string }
-}
-
-function extractGraphqlErrors(stderr: string, stdout: string): GhGraphqlErrorShape[] {
-  // `gh api graphql` prints the response JSON to stdout even on GraphQL
-  // errors, and the stderr carries a summary. Try stdout first; if parsing
-  // fails, fall back to stderr.
-  const sources = [stdout, stderr]
-  for (const src of sources) {
-    if (!src) continue
-    try {
-      const parsed = JSON.parse(src) as { errors?: GhGraphqlErrorShape[] }
-      if (parsed.errors && parsed.errors.length > 0) {
-        return parsed.errors
-      }
-    } catch {
-      // not JSON — continue
-    }
-  }
-  return []
-}
-
-function errorsIndicateParentField(errors: GhGraphqlErrorShape[], stderr: string): boolean {
-  const lower = stderr.toLowerCase()
-  // Preview-header shape: gh returns a 4xx with "preview" in the message.
-  if (lower.includes('preview') && lower.includes('parent')) return true
-  return errors.some((e) => {
-    const type = (e.type ?? '').toUpperCase()
-    if (type === 'FIELD_NOT_FOUND' || type === 'UNDEFINED_FIELD' || type === 'FIELD_ERRORS') {
-      const tail = e.path?.[e.path.length - 1]
-      if (tail === 'parent') return true
-      // FIELD_ERRORS often omits `path`; match on message for the parent field.
-      if ((e.message ?? '').toLowerCase().includes('parent')) return true
-    }
-    return false
-  })
-}
-
-export function classifyProjectError(stderr: string, stdout: string): GitHubProjectViewError {
-  const errors = extractGraphqlErrors(stderr, stdout)
-  const s = stderr.toLowerCase()
-
-  // Auth
-  if (s.includes('authentication required') || s.includes('not logged in') || s.includes('gh auth login')) {
-    return {
-      type: 'auth_required',
-      message: 'Sign in to GitHub to load project tasks. Run `gh auth login`.'
-    }
-  }
-  // Scope
-  if (
-    s.includes('missing required scope') ||
-    s.includes("your token has not been granted") ||
-    (s.includes('resource not accessible') && (s.includes('project') || s.includes('scope')))
-  ) {
-    return {
-      type: 'scope_missing',
-      message:
-        'GitHub project access needs additional scopes. Run `gh auth refresh -s project -s read:org -s repo`.'
-    }
-  }
-  // Rate limit
-  if (s.includes('rate limit') || s.includes('api rate limit exceeded')) {
-    return { type: 'rate_limited', message: 'GitHub rate limit hit. Try again in a few minutes.' }
-  }
-  // Network — checked BEFORE not_found because DNS failures surface as
-  // "could not resolve host", which would otherwise be partially matched by
-  // the not_found branch's "could not resolve" check. Substring matching here
-  // is a one-way trapdoor: a real GraphQL "Could not resolve to a User…"
-  // error always contains "to a", so we tighten the not_found check below to
-  // require that token.
-  if (
-    s.includes('timeout') ||
-    s.includes('no such host') ||
-    s.includes('network') ||
-    s.includes('could not resolve host') ||
-    s.includes('dial tcp')
-  ) {
-    return { type: 'network_error', message: 'Network error — check your connection.' }
-  }
-  // Not found
-  if (
-    s.includes('http 404') ||
-    errors.some((e) => (e.type ?? '').toUpperCase() === 'NOT_FOUND') ||
-    s.includes('could not resolve to a ')
-  ) {
-    const firstNotFound = errors.find((e) => (e.type ?? '').toUpperCase() === 'NOT_FOUND')
-    return {
-      type: 'not_found',
-      message: 'Project or view not found.',
-      details: firstNotFound
-        ? { path: firstNotFound.path, code: firstNotFound.extensions?.code }
-        : undefined
-    }
-  }
-  // Validation
-  if (s.includes('http 422') || s.includes('validation failed')) {
-    return { type: 'validation_error', message: `Invalid request — ${stderr.trim()}` }
-  }
-  // GraphQL error with structured info
-  if (errors.length > 0) {
-    const first = errors[0]
-    return {
-      type: 'unknown',
-      message: first.message ?? 'Unknown GraphQL error.',
-      details: { path: first.path, code: first.extensions?.code }
-    }
-  }
-  return { type: 'unknown', message: `GitHub request failed: ${stderr.trim()}` }
-}
-
-function driftError(
-  reason: string,
-  details?: { path?: Array<string | number>; code?: string }
-): GitHubProjectViewError {
-  return { type: 'schema_drift', message: `Could not read this project view: ${reason}.`, details }
-}
-
-// Why: the rate-limit circuit breaker short-circuits before we spawn `gh`
-// when the cached snapshot says we're below the safety floor. Synthesize the
-// same `rate_limited` error shape as the post-hoc classifier so the UI path
-// is unchanged. We DO NOT fail open here when there's no cached snapshot —
-// rateLimitGuard already handles that case (returns `blocked:false`).
-function rateLimitedError(
-  blocked: { remaining: number; limit: number; resetAt: number }
-): GitHubProjectViewError {
-  const resetIn = Math.max(0, blocked.resetAt - Math.floor(Date.now() / 1000))
-  const mins = Math.ceil(resetIn / 60)
-  return {
-    type: 'rate_limited',
-    message: `GitHub rate limit nearly exhausted (${blocked.remaining}/${blocked.limit} left). Resets in ~${mins}m.`
-  }
-}
-
-// ─── Low-level gh api graphql invocation ───────────────────────────────
-
-type GraphqlVars = Record<string, string | number | boolean>
-
-async function runGraphql<T>(
-  query: string,
-  vars: GraphqlVars,
-  cwd?: string
-): Promise<
-  | { ok: true; data: T }
-  | { ok: false; error: GitHubProjectViewError; raw: { stderr: string; stdout: string } }
-> {
-  const guard = rateLimitGuard('graphql')
-  if (guard.blocked) {
-    return { ok: false, error: rateLimitedError(guard), raw: { stderr: '', stdout: '' } }
-  }
-  // Why: build argv as an array. `-f` for strings (including numbers passed
-  // as strings), `-F` coerces to typed. We use `-f` uniformly and coerce in
-  // the query via Int! casts, because `gh` can confuse empty strings.
-  const args: string[] = ['api', 'graphql', '-f', `query=${query}`]
-  for (const [k, v] of Object.entries(vars)) {
-    if (typeof v === 'number' || typeof v === 'boolean') {
-      args.push('-F', `${k}=${String(v)}`)
-    } else {
-      args.push('-f', `${k}=${v}`)
-    }
-  }
-  await acquire()
-  noteRateLimitSpend('graphql')
-  try {
-    const { stdout, stderr } = await ghExecFileAsync(args, {
-      encoding: 'utf-8',
-      ...(cwd ? { cwd } : {})
-    })
-    try {
-      const parsed = JSON.parse(stdout) as { data?: T; errors?: GhGraphqlErrorShape[] }
-      if (parsed.errors && parsed.errors.length > 0) {
-        return {
-          ok: false,
-          error: classifyProjectError(stderr, stdout),
-          raw: { stderr, stdout }
-        }
-      }
-      if (parsed.data === undefined) {
-        return {
-          ok: false,
-          error: driftError('response missing data'),
-          raw: { stderr, stdout }
-        }
-      }
-      return { ok: true, data: parsed.data }
-    } catch (parseErr) {
-      return {
-        ok: false,
-        error: driftError(
-          `failed to parse response (${parseErr instanceof Error ? parseErr.message : String(parseErr)})`
-        ),
-        raw: { stderr, stdout }
-      }
-    }
-  } catch (err) {
-    // gh executable failures (non-zero exit). Read stderr/stdout from the
-    // exec rejection's explicit fields — `err.message` may truncate stderr.
-    const { stderr, stdout: maybeStdout } = extractExecError(err)
-    return {
-      ok: false,
-      error: classifyProjectError(stderr, maybeStdout),
-      raw: { stderr, stdout: maybeStdout }
-    }
-  } finally {
-    release()
-  }
-}
-
-async function runRest<T>(
-  args: string[],
-  cwd?: string,
-  bucket: RateLimitBucketKind = 'core',
-  options?: { expectEmpty?: boolean }
-): Promise<{ ok: true; data: T } | { ok: false; error: GitHubProjectViewError }> {
-  const guard = rateLimitGuard(bucket)
-  if (guard.blocked) {
-    return { ok: false, error: rateLimitedError(guard) }
-  }
-  await acquire()
-  noteRateLimitSpend(bucket)
-  try {
-    const { stdout, stderr } = await ghExecFileAsync(['api', ...args], {
-      encoding: 'utf-8',
-      ...(cwd ? { cwd } : {})
-    })
-    // Why: 204/empty-body endpoints (DELETE label, DELETE comment) return no
-    // body. Treat empty stdout as success rather than misclassifying the
-    // unparseable response as 'unknown' — which the caller would otherwise
-    // need to special-case and risks masking real failures whose stderr the
-    // classifier also tags as 'unknown'.
-    if (options?.expectEmpty && stdout.trim() === '') {
-      return { ok: true, data: undefined as T }
-    }
-    try {
-      return { ok: true, data: JSON.parse(stdout) as T }
-    } catch {
-      return {
-        ok: false,
-        error: { type: 'unknown', message: `Unexpected REST response: ${stderr.trim()}` }
-      }
-    }
-  } catch (err) {
-    const { stderr, stdout: maybeStdout } = extractExecError(err)
-    return { ok: false, error: classifyProjectError(stderr, maybeStdout) }
-  } finally {
-    release()
-  }
-}
-
 // ─── Normalizers ───────────────────────────────────────────────────────
 
 type RawProjectV2Field = {
@@ -423,15 +120,15 @@ type RawProjectV2Field = {
   id?: string
   name?: string
   dataType?: string
-  options?: Array<{ id?: string; name?: string; color?: string }>
+  options?: { id?: string; name?: string; color?: string }[]
   configuration?: {
-    iterations?: Array<{ id?: string; title?: string; startDate?: string; duration?: number }>
-    completedIterations?: Array<{
+    iterations?: { id?: string; title?: string; startDate?: string; duration?: number }[]
+    completedIterations?: {
       id?: string
       title?: string
       startDate?: string
       duration?: number
-    }>
+    }[]
   }
 }
 
@@ -487,7 +184,7 @@ type RawUser = {
 }
 
 function normalizeUser(raw: RawUser | null | undefined): GitHubProjectUser | null {
-  if (!raw || typeof raw.login !== 'string') return null
+  if (!raw || typeof raw.login !== 'string') {return null}
   return {
     login: raw.login,
     name: raw.name ?? null,
@@ -498,7 +195,7 @@ function normalizeUser(raw: RawUser | null | undefined): GitHubProjectUser | nul
 type RawLabel = { name?: string; color?: string }
 
 function normalizeLabel(raw: RawLabel | null | undefined): GitHubProjectLabel | null {
-  if (!raw || typeof raw.name !== 'string') return null
+  if (!raw || typeof raw.name !== 'string') {return null}
   return { name: raw.name, color: raw.color ?? '' }
 }
 
@@ -520,11 +217,11 @@ type RawFieldValue = {
 }
 
 export function normalizeFieldValue(raw: RawFieldValue | null | undefined): GitHubProjectFieldValue | null {
-  if (!raw || !raw.field || typeof raw.field.id !== 'string') return null
+  if (!raw || !raw.field || typeof raw.field.id !== 'string') {return null}
   const fieldId = raw.field.id
   switch (raw.__typename) {
     case 'ProjectV2ItemFieldSingleSelectValue':
-      if (typeof raw.optionId !== 'string') return null
+      if (typeof raw.optionId !== 'string') {return null}
       return {
         kind: 'single-select',
         fieldId,
@@ -533,7 +230,7 @@ export function normalizeFieldValue(raw: RawFieldValue | null | undefined): GitH
         color: raw.color ?? ''
       }
     case 'ProjectV2ItemFieldIterationValue':
-      if (typeof raw.iterationId !== 'string') return null
+      if (typeof raw.iterationId !== 'string') {return null}
       return {
         kind: 'iteration',
         fieldId,
@@ -545,7 +242,7 @@ export function normalizeFieldValue(raw: RawFieldValue | null | undefined): GitH
     case 'ProjectV2ItemFieldTextValue':
       return { kind: 'text', fieldId, text: raw.text ?? '' }
     case 'ProjectV2ItemFieldNumberValue':
-      if (typeof raw.number !== 'number') return null
+      if (typeof raw.number !== 'number') {return null}
       return { kind: 'number', fieldId, number: raw.number }
     case 'ProjectV2ItemFieldDateValue':
       return { kind: 'date', fieldId, date: raw.date ?? '' }
@@ -601,10 +298,10 @@ type NormalizedItemOutcome =
   | { ok: false; drift: GitHubProjectViewError }
 
 function mapItemType(raw: string | undefined, hasContent: boolean): GitHubProjectRowItemType {
-  if (raw === 'ISSUE') return 'ISSUE'
-  if (raw === 'PULL_REQUEST') return 'PULL_REQUEST'
-  if (raw === 'DRAFT_ISSUE') return 'DRAFT_ISSUE'
-  if (raw === 'REDACTED' || !hasContent) return 'REDACTED'
+  if (raw === 'ISSUE') {return 'ISSUE'}
+  if (raw === 'PULL_REQUEST') {return 'PULL_REQUEST'}
+  if (raw === 'DRAFT_ISSUE') {return 'DRAFT_ISSUE'}
+  if (raw === 'REDACTED' || !hasContent) {return 'REDACTED'}
   // Unknown item type with content — treat as redacted rather than dropping.
   return 'REDACTED'
 }
@@ -771,7 +468,7 @@ type RawProjectConfig = {
   url?: string
   views?: {
     pageInfo?: { hasNextPage?: boolean; endCursor?: string | null }
-    nodes?: Array<RawProjectView | null>
+    nodes?: (RawProjectView | null)[]
   }
 }
 
@@ -783,11 +480,11 @@ type RawProjectView = {
   filter?: string | null
   fields?: {
     pageInfo?: { hasNextPage?: boolean; endCursor?: string | null }
-    nodes?: Array<RawProjectV2Field | null>
+    nodes?: (RawProjectV2Field | null)[]
   }
-  groupByFields?: { nodes?: Array<RawProjectV2Field | null> }
+  groupByFields?: { nodes?: (RawProjectV2Field | null)[] }
   sortByFields?: {
-    nodes?: Array<{ direction?: string; field?: RawProjectV2Field | null } | null>
+    nodes?: ({ direction?: string; field?: RawProjectV2Field | null } | null)[]
   }
 }
 
@@ -838,12 +535,12 @@ async function fetchProjectViewsPage(args: {
     ${FIELD_CONFIG_FRAGMENT}
   `
   const vars: GraphqlVars = { owner: args.owner, num: args.projectNumber }
-  if (args.after) vars.after = args.after
+  if (args.after) {vars.after = args.after}
   const res = await runGraphql<Record<string, { projectV2?: RawProjectConfig | null } | null>>(
     query,
     vars
   )
-  if (!res.ok) return res
+  if (!res.ok) {return res}
   const top = res.data[root]
   const project = top?.projectV2 ?? null
   if (!project || typeof project.id !== 'string') {
@@ -893,11 +590,11 @@ async function fetchViewFieldsContinuation(
         id?: string
         fields?: {
           pageInfo?: { hasNextPage?: boolean; endCursor?: string | null }
-          nodes?: Array<RawProjectV2Field | null>
+          nodes?: (RawProjectV2Field | null)[]
         }
       } | null
     }>(query, { viewId, after: cursor })
-    if (!res.ok) return res
+    if (!res.ok) {return res}
     const view = res.data.node ?? null
     if (!view) {
       return { ok: false, error: driftError('view disappeared during field pagination') }
@@ -924,18 +621,18 @@ function finalizeView(
   const all = [...(raw.fields?.nodes ?? []), ...extraFields.map((f) => f as RawProjectV2Field)]
   for (const f of all) {
     const n = normalizeField(f)
-    if (n) fields.push(n)
+    if (n) {fields.push(n)}
   }
   const groupByFields: GitHubProjectField[] = []
   for (const f of raw.groupByFields?.nodes ?? []) {
     const n = normalizeField(f)
-    if (n) groupByFields.push(n)
+    if (n) {groupByFields.push(n)}
   }
   const sortByFields: GitHubProjectSort[] = []
   for (const s of raw.sortByFields?.nodes ?? []) {
-    if (!s || (s.direction !== 'ASC' && s.direction !== 'DESC')) continue
+    if (!s || (s.direction !== 'ASC' && s.direction !== 'DESC')) {continue}
     const n = normalizeField(s.field)
-    if (n) sortByFields.push({ direction: s.direction, field: n })
+    if (n) {sortByFields.push({ direction: s.direction, field: n })}
   }
   return {
     ok: true,
@@ -959,9 +656,9 @@ function matchesSelector(
   raw: RawProjectView,
   sel: { viewId?: string; viewNumber?: number; viewName?: string }
 ): 'none' | 'id' | 'number' | 'name' | 'default' {
-  if (sel.viewId && raw.id === sel.viewId) return 'id'
-  if (sel.viewNumber !== undefined && raw.number === sel.viewNumber) return 'number'
-  if (sel.viewName && raw.name === sel.viewName) return 'name'
+  if (sel.viewId && raw.id === sel.viewId) {return 'id'}
+  if (sel.viewNumber !== undefined && raw.number === sel.viewNumber) {return 'number'}
+  if (sel.viewName && raw.name === sel.viewName) {return 'name'}
   if (
     sel.viewId === undefined &&
     sel.viewNumber === undefined &&
@@ -978,7 +675,7 @@ function matchesSelector(
 type RawItemsPage = {
   totalCount?: number
   pageInfo?: { hasNextPage?: boolean; endCursor?: string | null }
-  nodes?: Array<RawItem | null>
+  nodes?: (RawItem | null)[]
 }
 
 // Why: runGraphql returns the classified error but not the raw GraphQL
@@ -1029,7 +726,7 @@ async function fetchItemsPageWithRaw(args: {
   argsArr.push('-F', `num=${args.projectNumber}`)
   argsArr.push('-f', `q=${args.query}`)
   argsArr.push('-F', `first=${args.first}`)
-  if (args.after) argsArr.push('-f', `after=${args.after}`)
+  if (args.after) {argsArr.push('-f', `after=${args.after}`)}
 
   const guard = rateLimitGuard('graphql')
   if (guard.blocked) {
@@ -1124,34 +821,42 @@ async function fetchAllItems(args: {
 > {
   // Why: if another caller is currently probing whether Issue.parent is
   // supported, await its decision so we don't fire a duplicate with-parent
-  // probe and don't capture a stale includeParent.
+  // probe and don't capture a stale includeParent. We must re-read
+  // parentFieldRetried AFTER awaiting because the probe may have flipped it.
   if (parentFieldProbeInFlight) {
     await parentFieldProbeInFlight.catch(() => {})
   }
   let includeParent = !parentFieldRetried
   let parentFieldDropped = parentFieldRetried
   // First page — single-flight the with-parent attempt so concurrent callers
-  // observe one probe result instead of each issuing their own.
+  // observe one probe result instead of each issuing their own. We must
+  // assign parentFieldProbeInFlight synchronously (no await between the
+  // null check and the assignment) so concurrent callers race on the same
+  // promise rather than each creating a duplicate probe.
   let first: Awaited<ReturnType<typeof fetchItemsPageWithRaw>>
+  let probePromise: Promise<Awaited<ReturnType<typeof fetchItemsPageWithRaw>>> | null = null
   if (includeParent && !parentFieldProbeInFlight) {
     let resolveProbe: () => void = () => {}
     parentFieldProbeInFlight = new Promise<void>((resolve) => {
       resolveProbe = resolve
     })
-    try {
-      first = await fetchItemsPageWithRaw({
-        owner: args.owner,
-        ownerType: args.ownerType,
-        projectNumber: args.projectNumber,
-        query: args.query,
-        first: ITEM_PAGE_SIZE,
-        after: null,
-        includeParent: true
-      })
-    } finally {
-      resolveProbe()
-      parentFieldProbeInFlight = null
-    }
+    probePromise = (async () => {
+      try {
+        return await fetchItemsPageWithRaw({
+          owner: args.owner,
+          ownerType: args.ownerType,
+          projectNumber: args.projectNumber,
+          query: args.query,
+          first: ITEM_PAGE_SIZE,
+          after: null,
+          includeParent: true
+        })
+      } finally {
+        resolveProbe()
+        parentFieldProbeInFlight = null
+      }
+    })()
+    first = await probePromise
   } else {
     first = await fetchItemsPageWithRaw({
       owner: args.owner,
@@ -1185,7 +890,7 @@ async function fetchAllItems(args: {
       includeParent: false
     })
   }
-  if (!first.ok) return { ok: false, error: first.error }
+  if (!first.ok) {return { ok: false, error: first.error }}
 
   // Drift guards
   if (first.page.totalCount === undefined || first.page.totalCount === null) {
@@ -1206,18 +911,18 @@ async function fetchAllItems(args: {
 
   const rows: GitHubProjectRow[] = []
   let position = 0
-  const appendNodes = (nodes: Array<RawItem | null>): GitHubProjectViewError | null => {
+  const appendNodes = (nodes: (RawItem | null)[]): GitHubProjectViewError | null => {
     for (const n of nodes) {
-      if (!n) continue
+      if (!n) {continue}
       const norm = normalizeItem(n, position)
-      if (!norm.ok) return norm.drift
+      if (!norm.ok) {return norm.drift}
       rows.push(norm.row)
       position++
     }
     return null
   }
   const e1 = appendNodes(first.page.nodes)
-  if (e1) return { ok: false, error: e1, totalCount }
+  if (e1) {return { ok: false, error: e1, totalCount }}
 
   // Paginate
   let hasNext = first.page.pageInfo.hasNextPage === true
@@ -1239,7 +944,7 @@ async function fetchAllItems(args: {
       after: cursor as string,
       includeParent
     })
-    if (!next.ok) return { ok: false, error: next.error, totalCount }
+    if (!next.ok) {return { ok: false, error: next.error, totalCount }}
     if (!Array.isArray(next.page.nodes)) {
       return { ok: false, error: driftError('items.nodes missing on follow page'), totalCount }
     }
@@ -1251,7 +956,7 @@ async function fetchAllItems(args: {
       }
     }
     const e2 = appendNodes(next.page.nodes)
-    if (e2) return { ok: false, error: e2, totalCount }
+    if (e2) {return { ok: false, error: e2, totalCount }}
     hasNext = next.page.pageInfo.hasNextPage === true
     cursor = next.page.pageInfo.endCursor
     if (hasNext && typeof cursor !== 'string') {
@@ -1286,7 +991,7 @@ async function fetchItemsCountOnly(args: {
   const res = await runGraphql<
     Record<string, { projectV2?: { items?: { totalCount?: number } | null } | null } | null>
   >(query, { owner: args.owner, num: args.projectNumber, q: args.query })
-  if (!res.ok) return null
+  if (!res.ok) {return null}
   const count = res.data[root]?.projectV2?.items?.totalCount
   return typeof count === 'number' ? count : null
 }
@@ -1297,9 +1002,9 @@ export async function getProjectViewTable(
   args: GetProjectViewTableArgs
 ): Promise<GetProjectViewTableResult> {
   const ownerCheck = assertSlug(args.owner, 'owner')
-  if (!ownerCheck.ok) return { ok: false, error: ownerCheck.error }
+  if (!ownerCheck.ok) {return { ok: false, error: ownerCheck.error }}
   const numCheck = assertPositiveInt(args.projectNumber, 'projectNumber')
-  if (!numCheck.ok) return { ok: false, error: numCheck.error }
+  if (!numCheck.ok) {return { ok: false, error: numCheck.error }}
   if (args.ownerType !== 'organization' && args.ownerType !== 'user') {
     return {
       ok: false,
@@ -1320,7 +1025,7 @@ export async function getProjectViewTable(
       projectNumber: args.projectNumber,
       after: cursor
     })
-    if (!page.ok) return { ok: false, error: page.error }
+    if (!page.ok) {return { ok: false, error: page.error }}
     project = page.project
     for (const v of page.views) {
       viewsSeen.push(v)
@@ -1329,7 +1034,7 @@ export async function getProjectViewTable(
         viewNumber: args.viewNumber,
         viewName: args.viewName
       })
-      if (m === 'none') continue
+      if (m === 'none') {continue}
       // Precedence: id > number > name > default.
       const rank: Record<typeof m, number> = { id: 4, number: 3, name: 2, default: 1 }
       const currentRank = matchStrength ? rank[matchStrength] : 0
@@ -1348,10 +1053,10 @@ export async function getProjectViewTable(
     // selector promotes a 'default' to a stronger match within the same
     // selector input — those ranks only matter when the caller supplied
     // a selector. Bail early on any non-null selectedRaw.
-    if (selectedRaw) break
-    if (!page.hasNextPage) break
+    if (selectedRaw) {break}
+    if (!page.hasNextPage) {break}
     cursor = page.endCursor
-    if (typeof cursor !== 'string') break
+    if (typeof cursor !== 'string') {break}
   }
   if (!project) {
     return { ok: false, error: { type: 'not_found', message: 'Project not found.' } }
@@ -1365,12 +1070,12 @@ export async function getProjectViewTable(
   const fieldsPi = selectedRaw.fields?.pageInfo
   if (fieldsPi?.hasNextPage === true && typeof fieldsPi.endCursor === 'string' && selectedRaw.id) {
     const cont = await fetchViewFieldsContinuation(selectedRaw.id, fieldsPi.endCursor)
-    if (!cont.ok) return { ok: false, error: cont.error }
+    if (!cont.ok) {return { ok: false, error: cont.error }}
     extraFields = cont.fields
   }
 
   const finalized = finalizeView(selectedRaw, extraFields)
-  if (!finalized.ok) return { ok: false, error: finalized.drift }
+  if (!finalized.ok) {return { ok: false, error: finalized.drift }}
   const selectedView = finalized.view
 
   // Why: an explicit empty-string override means "no filter"; treat undefined
@@ -1438,23 +1143,23 @@ type RawViewerDiscovery = {
     login?: string
     projectsV2?: {
       pageInfo?: { hasNextPage?: boolean; endCursor?: string | null }
-      nodes?: Array<{
+      nodes?: ({
         id?: string
         number?: number
         title?: string
         url?: string
         owner?: { __typename?: string; login?: string }
-      } | null>
+      } | null)[]
     }
     organizations?: {
       pageInfo?: { hasNextPage?: boolean; endCursor?: string | null }
-      nodes?: Array<{
+      nodes?: ({
         login?: string
         projectsV2?: {
           pageInfo?: { hasNextPage?: boolean; endCursor?: string | null }
           nodes?: Array<{ id?: string; number?: number; title?: string; url?: string } | null>
         }
-      } | null>
+      } | null)[]
     }
   }
 }
@@ -1491,7 +1196,7 @@ export async function listAccessibleProjects(): Promise<ListAccessibleProjectsRe
       }
     `
     const vars: GraphqlVars = {}
-    if (viewerCursor) vars.after = viewerCursor
+    if (viewerCursor) {vars.after = viewerCursor}
     const res = await runGraphql<RawViewerDiscovery>(query, vars)
     if (!res.ok) {
       // Why: a viewer-level failure is structural — if we can't list the
@@ -1503,10 +1208,10 @@ export async function listAccessibleProjects(): Promise<ListAccessibleProjectsRe
     if (!res.data.viewer) {
       return { ok: false, error: driftError('viewer missing') }
     }
-    if (viewerLogin === null) viewerLogin = res.data.viewer.login ?? null
+    if (viewerLogin === null) {viewerLogin = res.data.viewer.login ?? null}
     const nodes = res.data.viewer.projectsV2?.nodes ?? []
     for (const n of nodes) {
-      if (!n || typeof n.id !== 'string' || typeof n.number !== 'number') continue
+      if (!n || typeof n.id !== 'string' || typeof n.number !== 'number') {continue}
       const ownerLogin = n.owner?.login ?? viewerLogin ?? ''
       const ownerType: GitHubProjectOwnerType =
         n.owner?.__typename === 'Organization' ? 'organization' : 'user'
@@ -1520,7 +1225,7 @@ export async function listAccessibleProjects(): Promise<ListAccessibleProjectsRe
         source: 'viewer'
       })
       viewerFetched++
-      if (viewerFetched >= DISCOVERY_PROJECTS_PER_OWNER) break
+      if (viewerFetched >= DISCOVERY_PROJECTS_PER_OWNER) {break}
     }
     const pi = res.data.viewer.projectsV2?.pageInfo
     viewerMore = pi?.hasNextPage === true && typeof pi.endCursor === 'string'
@@ -1558,7 +1263,7 @@ export async function listAccessibleProjects(): Promise<ListAccessibleProjectsRe
       }
     `
     const vars: GraphqlVars = {}
-    if (orgCursor) vars.orgAfter = orgCursor
+    if (orgCursor) {vars.orgAfter = orgCursor}
     const res = await runGraphql<RawViewerDiscovery>(query, vars)
     if (!res.ok) {
       // Why: the org-listing query itself failed (not a nested projectsV2).
@@ -1571,8 +1276,8 @@ export async function listAccessibleProjects(): Promise<ListAccessibleProjectsRe
     }
     const orgs = res.data.viewer?.organizations?.nodes ?? []
     for (const org of orgs) {
-      if (!org || typeof org.login !== 'string') continue
-      if (orgsSeen >= DISCOVERY_MAX_ORGS) break
+      if (!org || typeof org.login !== 'string') {continue}
+      if (orgsSeen >= DISCOVERY_MAX_ORGS) {break}
       orgsSeen++
       const login = org.login
       // Cache owner → ownerType for downstream paste/resolve even when the
@@ -1582,8 +1287,8 @@ export async function listAccessibleProjects(): Promise<ListAccessibleProjectsRe
       const nodes = org.projectsV2?.nodes ?? []
       let ownerCount = 0
       for (const n of nodes) {
-        if (!n || typeof n.id !== 'string' || typeof n.number !== 'number') continue
-        if (ownerCount >= DISCOVERY_PROJECTS_PER_OWNER) break
+        if (!n || typeof n.id !== 'string' || typeof n.number !== 'number') {continue}
+        if (ownerCount >= DISCOVERY_PROJECTS_PER_OWNER) {break}
         orgProjects.push({
           id: n.id,
           owner: login,
@@ -1601,7 +1306,7 @@ export async function listAccessibleProjects(): Promise<ListAccessibleProjectsRe
     orgCursor = orgMore ? (pi?.endCursor ?? null) : null
   }
 
-  if (viewerLogin) ownerTypeCache.set(viewerLogin, 'user')
+  if (viewerLogin) {ownerTypeCache.set(viewerLogin, 'user')}
 
   return {
     ok: true,
@@ -1619,15 +1324,15 @@ type ParsedPaste =
 
 export function parseProjectPaste(input: string): ParsedPaste | null {
   const trimmed = input.trim()
-  if (!trimmed) return null
+  if (!trimmed) {return null}
   // URL forms
   const urlRe = /^https?:\/\/github\.com\/(orgs|users)\/([^/]+)\/projects\/(\d+)(?:\/views\/(\d+))?/i
   const m = trimmed.match(urlRe)
   if (m) {
     const [, kindSeg, owner, nStr, vStr] = m
     const number = parseInt(nStr, 10)
-    if (!Number.isInteger(number) || number < 1) return null
-    if (!isValidOwnerSlug(owner)) return null
+    if (!Number.isInteger(number) || number < 1) {return null}
+    if (!isValidOwnerSlug(owner)) {return null}
     const viewNumber = vStr ? parseInt(vStr, 10) : undefined
     return {
       kind: kindSeg === 'orgs' ? 'org' : 'user',
@@ -1643,7 +1348,7 @@ export function parseProjectPaste(input: string): ParsedPaste | null {
   const sm = trimmed.match(shortRe)
   if (sm) {
     const number = parseInt(sm[2], 10)
-    if (!Number.isInteger(number) || number < 1) return null
+    if (!Number.isInteger(number) || number < 1) {return null}
     return { kind: 'bare', owner: sm[1], number }
   }
   return null
@@ -1676,13 +1381,13 @@ async function resolveOwnerType(
         }
       `
     const vars: GraphqlVars = { owner }
-    if (num) vars.num = num
+    if (num) {vars.num = num}
     const res = await runGraphql<
       Record<string, { projectV2?: { id?: string; title?: string } | null; login?: string } | null>
     >(query, vars)
-    if (!res.ok) return { ok: false, error: res.error }
+    if (!res.ok) {return { ok: false, error: res.error }}
     const top = res.data[root]
-    if (!top) return { ok: false, error: { type: 'not_found', message: 'Owner not found.' } }
+    if (!top) {return { ok: false, error: { type: 'not_found', message: 'Owner not found.' } }}
     if (num) {
       const p = top.projectV2
       if (!p || typeof p.id !== 'string') {
@@ -1748,7 +1453,7 @@ export async function resolveProjectRef(
     parsed.kind === 'org' ? 'organization' : parsed.kind === 'user' ? 'user' : null
   // Verify by fetching project title.
   const ownerRes = await resolveOwnerType(parsed.owner, preferred)
-  if (!ownerRes.ok) return { ok: false, error: ownerRes.error }
+  if (!ownerRes.ok) {return { ok: false, error: ownerRes.error }}
   const ownerType = ownerRes.ownerType
   const root = ownerQueryRoot(ownerType)
   const query = `
@@ -1759,7 +1464,7 @@ export async function resolveProjectRef(
   const res = await runGraphql<
     Record<string, { projectV2?: { id?: string; title?: string } | null } | null>
   >(query, { owner: parsed.owner, num: parsed.number })
-  if (!res.ok) return { ok: false, error: res.error }
+  if (!res.ok) {return { ok: false, error: res.error }}
   const p = res.data[root]?.projectV2
   if (!p || typeof p.id !== 'string') {
     return { ok: false, error: { type: 'not_found', message: 'Project not found.' } }
@@ -1769,7 +1474,13 @@ export async function resolveProjectRef(
     owner: parsed.owner,
     ownerType,
     number: parsed.number,
-    title: p.title ?? ''
+    title: p.title ?? '',
+    // Why: forward the parsed view number from /views/{n} URLs so the
+    // renderer can skip the view-pick step. parsed.kind === 'bare' has no
+    // viewNumber (owner/number shorthand carries no view).
+    ...(parsed.kind !== 'bare' && parsed.viewNumber !== undefined
+      ? { viewNumber: parsed.viewNumber }
+      : {})
   }
 }
 
@@ -1779,9 +1490,9 @@ export async function listProjectViews(
   args: ListProjectViewsArgs
 ): Promise<ListProjectViewsResult> {
   const ownerCheck = assertSlug(args.owner, 'owner')
-  if (!ownerCheck.ok) return { ok: false, error: ownerCheck.error }
+  if (!ownerCheck.ok) {return { ok: false, error: ownerCheck.error }}
   const numCheck = assertPositiveInt(args.projectNumber, 'projectNumber')
-  if (!numCheck.ok) return { ok: false, error: numCheck.error }
+  if (!numCheck.ok) {return { ok: false, error: numCheck.error }}
   if (args.ownerType !== 'organization' && args.ownerType !== 'user') {
     return { ok: false, error: { type: 'validation_error', message: 'Invalid ownerType.' } }
   }
@@ -1794,9 +1505,9 @@ export async function listProjectViews(
       projectNumber: args.projectNumber,
       after: cursor
     })
-    if (!page.ok) return { ok: false, error: page.error }
+    if (!page.ok) {return { ok: false, error: page.error }}
     for (const v of page.views) {
-      if (typeof v.id !== 'string' || typeof v.layout !== 'string') continue
+      if (typeof v.id !== 'string' || typeof v.layout !== 'string') {continue}
       summaries.push({
         id: v.id,
         number: typeof v.number === 'number' ? v.number : 0,
@@ -1804,692 +1515,9 @@ export async function listProjectViews(
         layout: v.layout as GitHubProjectViewLayout
       })
     }
-    if (!page.hasNextPage) break
+    if (!page.hasNextPage) {break}
     cursor = page.endCursor
-    if (typeof cursor !== 'string') break
+    if (typeof cursor !== 'string') {break}
   }
   return { ok: true, views: summaries }
-}
-
-// ─── Project field mutations ──────────────────────────────────────────
-
-class UnknownFieldMutationKindError extends Error {
-  constructor(kind: string) {
-    super(`Unknown project field mutation kind: ${kind}`)
-  }
-}
-
-function graphqlValueForFieldMutation(value: GitHubProjectFieldMutationValue): string {
-  // Serialize the value fragment for the GraphQL mutation. We use GraphQL
-  // variables for every dynamic piece, so here we only pick the variable name
-  // to reference per value kind.
-  switch (value.kind) {
-    case 'single-select':
-      return 'singleSelectOptionId: $value'
-    case 'iteration':
-      return 'iterationId: $value'
-    case 'text':
-      return 'text: $value'
-    case 'number':
-      return 'number: $value'
-    case 'date':
-      return 'date: $value'
-    default:
-      // Why: defensive default. If a new mutation kind is added to the type
-      // but not handled here, returning undefined would silently produce a
-      // broken GraphQL query. Throw so updateProjectItemFieldValue can map it
-      // to a validation_error rather than dispatching a malformed mutation.
-      throw new UnknownFieldMutationKindError((value as { kind: string }).kind)
-  }
-}
-
-function mutationValueVar(value: GitHubProjectFieldMutationValue): {
-  type: string
-  val: string | number
-} {
-  switch (value.kind) {
-    case 'single-select':
-      return { type: 'String!', val: value.optionId }
-    case 'iteration':
-      return { type: 'String!', val: value.iterationId }
-    case 'text':
-      return { type: 'String!', val: value.text }
-    case 'number':
-      return { type: 'Float!', val: value.number }
-    case 'date':
-      return { type: 'Date!', val: value.date }
-    default:
-      // Why: see graphqlValueForFieldMutation — surface unknown kinds loudly
-      // instead of returning undefined and dispatching an invalid mutation.
-      throw new UnknownFieldMutationKindError((value as { kind: string }).kind)
-  }
-}
-
-export async function updateProjectItemFieldValue(
-  args: UpdateProjectItemFieldArgs
-): Promise<GitHubProjectMutationResult> {
-  if (!args.projectId || !args.itemId || !args.fieldId) {
-    return { ok: false, error: { type: 'validation_error', message: 'Missing ids.' } }
-  }
-  let valFrag: string
-  let valVar: { type: string; val: string | number }
-  try {
-    valFrag = graphqlValueForFieldMutation(args.value)
-    valVar = mutationValueVar(args.value)
-  } catch (err) {
-    if (err instanceof UnknownFieldMutationKindError) {
-      return { ok: false, error: { type: 'validation_error', message: err.message } }
-    }
-    throw err
-  }
-  const query = `
-    mutation($projectId:ID!, $itemId:ID!, $fieldId:ID!, $value:${valVar.type}) {
-      updateProjectV2ItemFieldValue(input: {
-        projectId: $projectId
-        itemId: $itemId
-        fieldId: $fieldId
-        value: { ${valFrag} }
-      }) { projectV2Item { id } }
-    }
-  `
-  const vars: GraphqlVars = {
-    projectId: args.projectId,
-    itemId: args.itemId,
-    fieldId: args.fieldId,
-    value: valVar.val
-  }
-  const res = await runGraphql<unknown>(query, vars)
-  if (!res.ok) return { ok: false, error: res.error }
-  return { ok: true }
-}
-
-export async function clearProjectItemFieldValue(
-  args: ClearProjectItemFieldArgs
-): Promise<GitHubProjectMutationResult> {
-  if (!args.projectId || !args.itemId || !args.fieldId) {
-    return { ok: false, error: { type: 'validation_error', message: 'Missing ids.' } }
-  }
-  const query = `
-    mutation($projectId:ID!, $itemId:ID!, $fieldId:ID!) {
-      clearProjectV2ItemFieldValue(input: {
-        projectId: $projectId
-        itemId: $itemId
-        fieldId: $fieldId
-      }) { projectV2Item { id } }
-    }
-  `
-  const res = await runGraphql<unknown>(query, {
-    projectId: args.projectId,
-    itemId: args.itemId,
-    fieldId: args.fieldId
-  })
-  if (!res.ok) return { ok: false, error: res.error }
-  return { ok: true }
-}
-
-// ─── Slug-addressed issue/PR mutations ────────────────────────────────
-
-function validateSlugArgs(
-  owner: unknown,
-  repo: unknown
-): { ok: true } | { ok: false; error: GitHubProjectViewError } {
-  const o = assertSlug(owner, 'owner')
-  if (!o.ok) return { ok: false, error: o.error }
-  const r = assertSlug(repo, 'repo')
-  if (!r.ok) return { ok: false, error: r.error }
-  return { ok: true }
-}
-
-export async function updateIssueBySlug(
-  args: UpdateIssueBySlugArgs
-): Promise<GitHubProjectMutationResult> {
-  const v = validateSlugArgs(args.owner, args.repo)
-  if (!v.ok) return v
-  const n = assertPositiveInt(args.number, 'number')
-  if (!n.ok) return { ok: false, error: n.error }
-  if (!args.updates || typeof args.updates !== 'object') {
-    return { ok: false, error: { type: 'validation_error', message: 'Updates required.' } }
-  }
-  const { title, body, state, addLabels, removeLabels, addAssignees, removeAssignees } = args.updates
-
-  // Title / body / state go through PATCH /repos/{owner}/{repo}/issues/{n}.
-  // Labels/assignees go through their dedicated endpoints.
-  const base = `repos/${args.owner}/${args.repo}/issues/${args.number}`
-
-  // 1) PATCH body
-  if (title !== undefined || body !== undefined || state !== undefined) {
-    const patchArgs: string[] = ['-X', 'PATCH', base]
-    if (title !== undefined) patchArgs.push('--raw-field', `title=${title}`)
-    if (body !== undefined) patchArgs.push('--raw-field', `body=${body}`)
-    if (state !== undefined) patchArgs.push('--raw-field', `state=${state}`)
-    const r = await runRest<unknown>(patchArgs)
-    if (!r.ok) return { ok: false, error: r.error }
-  }
-
-  // 2) Labels — collapse multi-delete fan-out into a single PUT when removing
-  //    >1 label. PUT /labels replaces the entire label set, so we fetch the
-  //    current labels first and compute the resulting set client-side. This
-  //    turns an N-delete + 1-add (=N+1 calls) into 1-fetch + 1-PUT (=2 calls)
-  //    once removeLabels has more than one entry, capping the cost at 2 even
-  //    for a "remove all 20 labels" mutation.
-  const removeCount = removeLabels?.length ?? 0
-  const addCount = addLabels?.length ?? 0
-  if (removeCount > 1) {
-    type RawLabelResp = { name?: string }[]
-    const fetched = await runRest<RawLabelResp>(['-X', 'GET', `${base}/labels`])
-    if (!fetched.ok) return { ok: false, error: fetched.error }
-    const currentNames = new Set(
-      fetched.data.map((l) => l.name).filter((n): n is string => typeof n === 'string')
-    )
-    for (const l of removeLabels ?? []) currentNames.delete(l)
-    for (const l of addLabels ?? []) currentNames.add(l)
-    const putArgs = ['-X', 'PUT', `${base}/labels`]
-    for (const name of currentNames) putArgs.push('--raw-field', `labels[]=${name}`)
-    // Why: PUT with empty `labels[]` clears all labels — that's the desired
-    // outcome when the user removed every label. gh requires `--raw-field` to
-    // construct the request body; an empty form is a valid clear-all.
-    const r = await runRest<unknown>(putArgs)
-    if (!r.ok) return { ok: false, error: r.error }
-  } else {
-    if (addCount > 0) {
-      const restArgs = ['-X', 'POST', `${base}/labels`]
-      for (const l of addLabels ?? []) restArgs.push('--raw-field', `labels[]=${l}`)
-      const r = await runRest<unknown>(restArgs)
-      if (!r.ok) return { ok: false, error: r.error }
-    }
-    if (removeCount === 1) {
-      const r = await runRest<unknown>(
-        ['-X', 'DELETE', `${base}/labels/${encodeURIComponent(removeLabels![0])}`],
-        undefined,
-        'core',
-        { expectEmpty: true }
-      )
-      if (!r.ok && r.error.type !== 'not_found') return { ok: false, error: r.error }
-    }
-  }
-
-  // 3) Assignees — POST and DELETE both accept arrays in a single call, so
-  //    add/remove are at most 2 calls regardless of array size.
-  if (addAssignees && addAssignees.length > 0) {
-    const restArgs = ['-X', 'POST', `${base}/assignees`]
-    for (const u of addAssignees) restArgs.push('--raw-field', `assignees[]=${u}`)
-    const r = await runRest<unknown>(restArgs)
-    if (!r.ok) return { ok: false, error: r.error }
-  }
-  if (removeAssignees && removeAssignees.length > 0) {
-    const restArgs = ['-X', 'DELETE', `${base}/assignees`]
-    for (const u of removeAssignees) restArgs.push('--raw-field', `assignees[]=${u}`)
-    const r = await runRest<unknown>(restArgs)
-    if (!r.ok) return { ok: false, error: r.error }
-  }
-  return { ok: true }
-}
-
-export async function updatePullRequestBySlug(
-  args: UpdatePullRequestBySlugArgs
-): Promise<GitHubProjectMutationResult> {
-  const v = validateSlugArgs(args.owner, args.repo)
-  if (!v.ok) return v
-  const n = assertPositiveInt(args.number, 'number')
-  if (!n.ok) return { ok: false, error: n.error }
-  if (!args.updates || typeof args.updates !== 'object') {
-    return { ok: false, error: { type: 'validation_error', message: 'Updates required.' } }
-  }
-  const patchArgs: string[] = ['-X', 'PATCH', `repos/${args.owner}/${args.repo}/pulls/${args.number}`]
-  // Why: count fields explicitly rather than inferring from patchArgs.length —
-  // adding a future header/flag arg silently breaks an array-length check.
-  let fieldCount = 0
-  if (args.updates.title !== undefined) {
-    patchArgs.push('--raw-field', `title=${args.updates.title}`)
-    fieldCount++
-  }
-  if (args.updates.body !== undefined) {
-    patchArgs.push('--raw-field', `body=${args.updates.body}`)
-    fieldCount++
-  }
-  if (fieldCount === 0) {
-    // No fields to update — nothing to do.
-    return { ok: true }
-  }
-  const r = await runRest<unknown>(patchArgs)
-  if (!r.ok) return { ok: false, error: r.error }
-  return { ok: true }
-}
-
-type RawIssueCommentResponse = {
-  id?: number
-  user?: { login?: string; avatar_url?: string; type?: string } | null
-  body?: string
-  created_at?: string
-  html_url?: string
-}
-
-function mapIssueComment(data: RawIssueCommentResponse, fallbackBody: string): PRComment {
-  return {
-    id: data.id ?? Date.now(),
-    author: data.user?.login ?? 'You',
-    authorAvatarUrl: data.user?.avatar_url ?? '',
-    body: data.body ?? fallbackBody,
-    createdAt: data.created_at ?? new Date().toISOString(),
-    url: data.html_url ?? '',
-    isBot: data.user?.type === 'Bot'
-  }
-}
-
-export async function addIssueCommentBySlug(
-  args: AddIssueCommentBySlugArgs
-): Promise<GitHubProjectCommentMutationResult> {
-  const v = validateSlugArgs(args.owner, args.repo)
-  if (!v.ok) return v
-  const n = assertPositiveInt(args.number, 'number')
-  if (!n.ok) return { ok: false, error: n.error }
-  if (typeof args.body !== 'string' || !args.body.trim()) {
-    return { ok: false, error: { type: 'validation_error', message: 'Comment body required.' } }
-  }
-  const r = await runRest<RawIssueCommentResponse>([
-    '-X',
-    'POST',
-    `repos/${args.owner}/${args.repo}/issues/${args.number}/comments`,
-    '--raw-field',
-    `body=${args.body}`
-  ])
-  if (!r.ok) return { ok: false, error: r.error }
-  return { ok: true, comment: mapIssueComment(r.data, args.body) }
-}
-
-export async function updateIssueCommentBySlug(
-  args: UpdateIssueCommentBySlugArgs
-): Promise<GitHubProjectMutationResult> {
-  const v = validateSlugArgs(args.owner, args.repo)
-  if (!v.ok) return v
-  const n = assertPositiveInt(args.commentId, 'commentId')
-  if (!n.ok) return { ok: false, error: n.error }
-  if (typeof args.body !== 'string' || !args.body.trim()) {
-    return { ok: false, error: { type: 'validation_error', message: 'Comment body required.' } }
-  }
-  const r = await runRest<unknown>([
-    '-X',
-    'PATCH',
-    `repos/${args.owner}/${args.repo}/issues/comments/${args.commentId}`,
-    '--raw-field',
-    `body=${args.body}`
-  ])
-  if (!r.ok) return { ok: false, error: r.error }
-  return { ok: true }
-}
-
-export async function deleteIssueCommentBySlug(
-  args: DeleteIssueCommentBySlugArgs
-): Promise<GitHubProjectMutationResult> {
-  const v = validateSlugArgs(args.owner, args.repo)
-  if (!v.ok) return v
-  const n = assertPositiveInt(args.commentId, 'commentId')
-  if (!n.ok) return { ok: false, error: n.error }
-  const r = await runRest<unknown>(
-    ['-X', 'DELETE', `repos/${args.owner}/${args.repo}/issues/comments/${args.commentId}`],
-    undefined,
-    'core',
-    { expectEmpty: true }
-  )
-  if (!r.ok) return { ok: false, error: r.error }
-  return { ok: true }
-}
-
-// ─── Slug-addressed picker sources ────────────────────────────────────
-
-export async function listLabelsBySlug(
-  args: ListLabelsBySlugArgs
-): Promise<ListLabelsBySlugResult> {
-  const v = validateSlugArgs(args.owner, args.repo)
-  if (!v.ok) return v
-  const guard = rateLimitGuard('core')
-  if (guard.blocked) return { ok: false, error: rateLimitedError(guard) }
-  await acquire()
-  // Why: `--paginate` may fan out to multiple pages; we can only reasonably
-  // estimate a 1-call spend up front. The next probe will reconcile.
-  noteRateLimitSpend('core')
-  try {
-    const { stdout } = await ghExecFileAsync(
-      ['api', '--paginate', `repos/${args.owner}/${args.repo}/labels`, '--jq', '.[].name'],
-      { encoding: 'utf-8' }
-    )
-    return {
-      ok: true,
-      labels: stdout
-        .trim()
-        .split('\n')
-        .filter((l) => l.length > 0)
-    }
-  } catch (err) {
-    const { stderr, stdout: maybeStdout } = extractExecError(err)
-    return { ok: false, error: classifyProjectError(stderr, maybeStdout) }
-  } finally {
-    release()
-  }
-}
-
-export async function listAssignableUsersBySlug(
-  args: ListAssignableUsersBySlugArgs
-): Promise<ListAssignableUsersBySlugResult> {
-  const v = validateSlugArgs(args.owner, args.repo)
-  if (!v.ok) return v
-  // Seed logins merge after the fetch so callers can include currently-visible
-  // assignees even if the repo participant search is sparse.
-  const result: GitHubAssignableUser[] = []
-  const guard = rateLimitGuard('core')
-  if (guard.blocked) return { ok: false, error: rateLimitedError(guard) }
-  await acquire()
-  noteRateLimitSpend('core')
-  try {
-    const { stdout } = await ghExecFileAsync(
-      [
-        'api',
-        '--paginate',
-        `repos/${args.owner}/${args.repo}/assignees`,
-        '--jq',
-        '.[] | {login: .login, name: null, avatarUrl: .avatar_url}'
-      ],
-      { encoding: 'utf-8' }
-    )
-    for (const line of stdout.trim().split('\n').filter((l) => l.length > 0)) {
-      try {
-        const u = JSON.parse(line) as { login?: string; avatarUrl?: string; name?: string | null }
-        if (typeof u.login === 'string') {
-          result.push({ login: u.login, name: u.name ?? null, avatarUrl: u.avatarUrl ?? '' })
-        }
-      } catch {
-        // skip malformed jq line
-      }
-    }
-  } catch (err) {
-    const { stderr } = extractExecError(err)
-    return { ok: false, error: classifyProjectError(stderr, '') }
-  } finally {
-    release()
-  }
-  if (args.seedLogins) {
-    const seen = new Set(result.map((u) => u.login))
-    for (const login of args.seedLogins) {
-      if (typeof login === 'string' && !seen.has(login)) {
-        result.push({ login, name: null, avatarUrl: '' })
-        seen.add(login)
-      }
-    }
-  }
-  return { ok: true, users: result }
-}
-
-// Why: Issue Types are a repo-level taxonomy (Bug/Feature/Task/etc) only
-// available on repos opted into typed-issues. Empty list (or schema_drift on
-// older GitHub deployments) is the legitimate "this repo doesn't use issue
-// types" signal — callers should treat it as "no editor".
-export async function listIssueTypesBySlug(
-  args: ListIssueTypesBySlugArgs
-): Promise<ListIssueTypesBySlugResult> {
-  const v = validateSlugArgs(args.owner, args.repo)
-  if (!v.ok) return v
-  const query = `
-    query($owner:String!, $repo:String!) {
-      repository(owner:$owner, name:$repo) {
-        issueTypes(first:50) {
-          nodes { id name color description }
-        }
-      }
-    }
-  `
-  const res = await runGraphql<{
-    repository?: {
-      issueTypes?: {
-        nodes?: Array<{
-          id?: string
-          name?: string
-          color?: string | null
-          description?: string | null
-        } | null>
-      } | null
-    } | null
-  }>(query, { owner: args.owner, repo: args.repo })
-  if (!res.ok) {
-    // Why: repos without issue types respond with a GraphQL error claiming the
-    // `issueTypes` field is unknown. Map that to an empty list so the UI shows
-    // "no editor" instead of an angry banner.
-    if (res.error.type === 'schema_drift' || res.error.type === 'validation_error') {
-      return { ok: true, types: [] }
-    }
-    return { ok: false, error: res.error }
-  }
-  const nodes = res.data.repository?.issueTypes?.nodes ?? []
-  const types = nodes
-    .filter((n): n is NonNullable<typeof n> => n !== null && typeof n.id === 'string' && typeof n.name === 'string')
-    .map((n) => ({
-      id: n.id as string,
-      name: n.name as string,
-      color: typeof n.color === 'string' ? n.color : null,
-      description: typeof n.description === 'string' ? n.description : null
-    }))
-  return { ok: true, types }
-}
-
-export async function updateIssueTypeBySlug(
-  args: UpdateIssueTypeBySlugArgs
-): Promise<GitHubProjectMutationResult> {
-  const v = validateSlugArgs(args.owner, args.repo)
-  if (!v.ok) return v
-  const n = assertPositiveInt(args.number, 'number')
-  if (!n.ok) return { ok: false, error: n.error }
-  // Why: `updateIssueIssueType` is the dedicated mutation; passing null for
-  // `issueTypeId` clears the type. We resolve the issue id via a lightweight
-  // GraphQL lookup because the REST endpoint doesn't accept issue types.
-  const lookup = await runGraphql<{
-    repository?: { issue?: { id?: string } | null } | null
-  }>(
-    `query($owner:String!, $repo:String!, $num:Int!) {
-       repository(owner:$owner, name:$repo) { issue(number:$num) { id } }
-     }`,
-    { owner: args.owner, repo: args.repo, num: args.number }
-  )
-  if (!lookup.ok) return { ok: false, error: lookup.error }
-  const issueId = lookup.data.repository?.issue?.id
-  if (!issueId) {
-    return { ok: false, error: { type: 'not_found', message: 'Issue not found.' } }
-  }
-  // Why: build the mutation conditionally so a null clear doesn't have to
-  // smuggle a null GraphQL variable through `gh api graphql -f`. The
-  // mutation accepts a literal `null` in the input object directly.
-  const query = args.issueTypeId
-    ? `
-        mutation($issueId:ID!, $issueTypeId:ID!) {
-          updateIssueIssueType(input: { issueId: $issueId, issueTypeId: $issueTypeId }) {
-            issue { id }
-          }
-        }
-      `
-    : `
-        mutation($issueId:ID!) {
-          updateIssueIssueType(input: { issueId: $issueId, issueTypeId: null }) {
-            issue { id }
-          }
-        }
-      `
-  const vars: GraphqlVars = args.issueTypeId
-    ? { issueId, issueTypeId: args.issueTypeId }
-    : { issueId }
-  const res = await runGraphql<unknown>(query, vars)
-  if (!res.ok) return { ok: false, error: res.error }
-  return { ok: true }
-}
-
-// ─── Slug-addressed work-item details ─────────────────────────────────
-
-export async function getWorkItemDetailsBySlug(
-  args: ProjectWorkItemDetailsBySlugArgs
-): Promise<ProjectWorkItemDetailsBySlugResult> {
-  const v = validateSlugArgs(args.owner, args.repo)
-  if (!v.ok) return v
-  const n = assertPositiveInt(args.number, 'number')
-  if (!n.ok) return { ok: false, error: n.error }
-  if (args.type !== 'issue' && args.type !== 'pr') {
-    return { ok: false, error: { type: 'validation_error', message: 'Invalid type.' } }
-  }
-
-  // Single GraphQL round-trip to fetch the issue/PR summary + comments + labels + assignees.
-  const contentFrag =
-    args.type === 'issue'
-      ? `
-        issue(number:$num) {
-          id number title url state stateReason updatedAt
-          body
-          author { login }
-          labels(first:50) { nodes { name } }
-          assignees(first:50) { nodes { login } }
-          participants(first:50) { nodes { login name avatarUrl } }
-          comments(first:100) {
-            nodes {
-              databaseId
-              author { login avatarUrl __typename }
-              body createdAt url
-            }
-          }
-        }
-      `
-      : `
-        pullRequest(number:$num) {
-          id number title url state isDraft updatedAt headRefName baseRefName
-          body
-          author { login }
-          labels(first:50) { nodes { name } }
-          assignees(first:50) { nodes { login } }
-          participants(first:50) { nodes { login name avatarUrl } }
-          comments(first:100) {
-            nodes {
-              databaseId
-              author { login avatarUrl __typename }
-              body createdAt url
-            }
-          }
-        }
-      `
-  const query = `
-    query($owner:String!, $repo:String!, $num:Int!) {
-      repository(owner:$owner, name:$repo) {
-        ${contentFrag}
-      }
-    }
-  `
-  const res = await runGraphql<{
-    repository?: {
-      issue?: RawContent & {
-        updatedAt?: string
-        body?: string
-        author?: { login?: string } | null
-        participants?: { nodes?: RawUser[] }
-        comments?: {
-          nodes?: Array<{
-            databaseId?: number
-            author?: { login?: string; avatarUrl?: string; __typename?: string } | null
-            body?: string
-            createdAt?: string
-            url?: string
-          } | null>
-        }
-      } | null
-      pullRequest?: RawContent & {
-        updatedAt?: string
-        body?: string
-        headRefName?: string
-        baseRefName?: string
-        author?: { login?: string } | null
-        participants?: { nodes?: RawUser[] }
-        comments?: {
-          nodes?: Array<{
-            databaseId?: number
-            author?: { login?: string; avatarUrl?: string; __typename?: string } | null
-            body?: string
-            createdAt?: string
-            url?: string
-          } | null>
-        }
-      } | null
-    } | null
-  }>(query, { owner: args.owner, repo: args.repo, num: args.number })
-  if (!res.ok) return { ok: false, error: res.error }
-  const raw = args.type === 'issue' ? res.data.repository?.issue : res.data.repository?.pullRequest
-  if (!raw) {
-    return { ok: false, error: { type: 'not_found', message: 'Item not found.' } }
-  }
-
-  const labels = (raw.labels?.nodes ?? [])
-    .map((l) => l?.name)
-    .filter((n): n is string => typeof n === 'string')
-  const assignees = (raw.assignees?.nodes ?? [])
-    .map((a) => a?.login)
-    .filter((l): l is string => typeof l === 'string')
-  const comments: PRComment[] = []
-  for (const c of raw.comments?.nodes ?? []) {
-    if (!c || typeof c.body !== 'string') continue
-    comments.push({
-      id: typeof c.databaseId === 'number' ? c.databaseId : Date.now(),
-      author: c.author?.login ?? '',
-      authorAvatarUrl: c.author?.avatarUrl ?? '',
-      body: c.body,
-      createdAt: typeof c.createdAt === 'string' ? c.createdAt : '',
-      url: typeof c.url === 'string' ? c.url : '',
-      isBot: c.author?.__typename === 'Bot'
-    })
-  }
-  const participants: GitHubAssignableUser[] = []
-  for (const p of raw.participants?.nodes ?? []) {
-    if (p && typeof p.login === 'string') {
-      participants.push({ login: p.login, name: p.name ?? null, avatarUrl: p.avatarUrl ?? '' })
-    }
-  }
-
-  const state: 'open' | 'closed' | 'merged' | 'draft' =
-    args.type === 'pr'
-      ? raw.isDraft
-        ? 'draft'
-        : raw.state === 'MERGED'
-          ? 'merged'
-          : raw.state === 'CLOSED'
-            ? 'closed'
-            : 'open'
-      : raw.state === 'CLOSED'
-        ? 'closed'
-        : 'open'
-
-  const details: GitHubWorkItemDetails = {
-    item: {
-      id: typeof raw.id === 'string' ? raw.id : '',
-      type: args.type,
-      number: typeof raw.number === 'number' ? raw.number : args.number,
-      title: typeof raw.title === 'string' ? raw.title : '',
-      state,
-      url: typeof raw.url === 'string' ? raw.url : '',
-      labels,
-      updatedAt:
-        typeof (raw as { updatedAt?: string }).updatedAt === 'string'
-          ? (raw as { updatedAt: string }).updatedAt
-          : '',
-      author:
-        typeof (raw as { author?: { login?: string } | null }).author?.login === 'string'
-          ? ((raw as { author: { login: string } }).author.login as string)
-          : null,
-      branchName:
-        args.type === 'pr' && typeof (raw as { headRefName?: string }).headRefName === 'string'
-          ? ((raw as { headRefName: string }).headRefName as string)
-          : undefined,
-      baseRefName:
-        args.type === 'pr' && typeof (raw as { baseRefName?: string }).baseRefName === 'string'
-          ? ((raw as { baseRefName: string }).baseRefName as string)
-          : undefined
-    },
-    body: typeof raw.body === 'string' ? raw.body : '',
-    comments,
-    participants,
-    // Why: PR files/checks/review-thread tabs depend on a local repo path and
-    // are out of Project-mode slug scope for v1. Omit them here; the dialog
-    // branches on their absence and hides those tabs.
-    ...(args.type === 'issue' ? { assignees } : {})
-  }
-  return { ok: true, details }
 }
