+++ src/renderer/src/components/github-project/ProjectViewList.tsx
@@ -71,17 +71,17 @@ export default function ProjectViewList({
   const toggleColumn = (fieldId: string): void => {
     setHidden((prev) => {
       const next = new Set(prev)
-      if (next.has(fieldId)) next.delete(fieldId)
-      else next.add(fieldId)
+      if (next.has(fieldId)) {next.delete(fieldId)}
+      else {next.add(fieldId)}
       saveHiddenColumns(scopeKey, next)
       return next
     })
   }
 
   const effectiveTable = useMemo<GitHubProjectTable>(() => {
-    if (!sortOverride) return table
+    if (!sortOverride) {return table}
     const field = fields.find((f) => f.id === sortOverride.fieldId)
-    if (!field) return table
+    if (!field) {return table}
     return {
       ...table,
       selectedView: {
@@ -101,8 +101,8 @@ export default function ProjectViewList({
 
   const handleSortClick = (fieldId: string): void => {
     setSortOverride((prev) => {
-      if (!prev || prev.fieldId !== fieldId) return { fieldId, direction: 'ASC' }
-      if (prev.direction === 'ASC') return { fieldId, direction: 'DESC' }
+      if (!prev || prev.fieldId !== fieldId) {return { fieldId, direction: 'ASC' }}
+      if (prev.direction === 'ASC') {return { fieldId, direction: 'DESC' }}
       return null
     })
   }
@@ -148,8 +148,8 @@ export default function ProjectViewList({
                 onToggle={() => {
                   setCollapsed((prev) => {
                     const next = new Set(prev)
-                    if (next.has(g.key)) next.delete(g.key)
-                    else next.add(g.key)
+                    if (next.has(g.key)) {next.delete(g.key)}
+                    else {next.add(g.key)}
                     return next
                   })
                 }}
