+++ src/renderer/src/components/github-project/group-sort.ts
@@ -113,8 +113,8 @@ export function groupRows(
   // Ordering rules per design doc §Grouping.
   entries.sort((a, b) => {
     // Empty group always last.
-    if (a[0] === EMPTY_GROUP_KEY) return 1
-    if (b[0] === EMPTY_GROUP_KEY) return -1
+    if (a[0] === EMPTY_GROUP_KEY) {return 1}
+    if (b[0] === EMPTY_GROUP_KEY) {return -1}
     if (groupField.kind === 'iteration' || groupField.kind === 'single-select') {
       return a[1].orderHint - b[1].orderHint
     }
@@ -133,9 +133,9 @@ function compareSort(a: GitHubProjectRow, b: GitHubProjectRow, sort: GitHubProje
   const aValue = a.fieldValuesByFieldId[field.id]
   const bValue = b.fieldValuesByFieldId[field.id]
   // Missing values sort last (regardless of direction).
-  if (!aValue && !bValue) return 0
-  if (!aValue) return 1
-  if (!bValue) return -1
+  if (!aValue && !bValue) {return 0}
+  if (!aValue) {return 1}
+  if (!bValue) {return -1}
 
   let cmp = 0
   if (field.kind === 'single-select' && aValue.kind === 'single-select' && bValue.kind === 'single-select') {
@@ -159,17 +159,17 @@ function compareSort(a: GitHubProjectRow, b: GitHubProjectRow, sort: GitHubProje
   } else if (aValue.kind === 'users' && bValue.kind === 'users') {
     const aLogin = aValue.users[0]?.login ?? ''
     const bLogin = bValue.users[0]?.login ?? ''
-    if (!aLogin && !bLogin) cmp = 0
-    else if (!aLogin) cmp = 1
-    else if (!bLogin) cmp = -1
-    else cmp = aLogin.localeCompare(bLogin)
+    if (!aLogin && !bLogin) {cmp = 0}
+    else if (!aLogin) {cmp = 1}
+    else if (!bLogin) {cmp = -1}
+    else {cmp = aLogin.localeCompare(bLogin)}
   } else if (aValue.kind === 'labels' && bValue.kind === 'labels') {
     const aName = aValue.labels[0]?.name ?? ''
     const bName = bValue.labels[0]?.name ?? ''
-    if (!aName && !bName) cmp = 0
-    else if (!aName) cmp = 1
-    else if (!bName) cmp = -1
-    else cmp = aName.localeCompare(bName)
+    if (!aName && !bName) {cmp = 0}
+    else if (!aName) {cmp = 1}
+    else if (!bName) {cmp = -1}
+    else {cmp = aName.localeCompare(bName)}
   } else {
     // Why: unknown sort-field kind — ignore this sort field and fall through
     // to tie-breaks (and eventually row.position). Dev-time warning gated so
@@ -192,7 +192,7 @@ export function sortRows(
   out.sort((a, b) => {
     for (const sort of sorts) {
       const cmp = compareSort(a, b, sort)
-      if (cmp !== 0) return cmp
+      if (cmp !== 0) {return cmp}
     }
     // Final tie-break: row.position preserves GitHub rank order.
     return a.position - b.position
@@ -206,7 +206,7 @@ export function isIterationCurrent(iteration: {
 }): boolean {
   // Parse as YYYY-MM-DD in UTC to avoid TZ-shift false negatives near midnight.
   const start = new Date(`${iteration.startDate}T00:00:00Z`).getTime()
-  if (Number.isNaN(start)) return false
+  if (Number.isNaN(start)) {return false}
   const end = start + iteration.duration * 86_400_000
   const now = Date.now()
   return now >= start && now < end
